Begin moving css files to central css folder.
