+++ client/src/components/App.js
@@ -6,7 +6,7 @@ import SuggestEdit from "./SuggestEdit";
 import Results from "./Results/Results";
 import Details from "./Details";
 import Nav from "./Navbar/Nav";
-import Footer from "./Footer/Footer";
+import Footer from "./Footer";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import {
   getNodeData,
