--- client/src/components/Home/EmergencyButton/EmergencyButton.css
@@ -1,57 +1,0 @@
-.emergency-button {
-  width: 95%;
-  max-width: 1000px;
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: .5em;
-  margin-top: 1em;
-  cursor: pointer;
-  background-color: #990000;
-  border: 2px solid white;
-  border-radius: 5px;
-}
-
-.emergency-button p {
-  color: white;
-  font-size: 1.5em;
-  line-height: 0;
-  font-family: "HelveticaNeue-CondensedBold", "Helvetica Neue Bold Condensed", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif; 
-  font-weight: 900;
-  font-stretch: condensed;
-}
-
-.emergency-button img {
-  max-height: 1.5em;
-  margin: 0 1em;
-  padding: 0;
-}
-
-@media (max-width: 600px) {
-  .emergency-button img {
-    display: none;
-  }
-  .emergency-button {
-    font-size: 2.25vh;
-  }
-} 
-
-@media (min-width: 768px) {
-  .emergency-button img {
-    padding: none;
-    margin: 0 2em;
-  }
-} 
-
-@media (min-width: 992px) {
-  .emergency-button {
-    width: 80%;
-    font-size: 1.5em;
-    margin-top: 0px;
-  }
-  .emergency-button img {
-    max-height: 1.5em;
-    padding: none;
-  }
-}
